+++ apps/chat/src/constants/apps-editor.ts
@@ -1,16 +1,3 @@
-export enum AppsEditorQuery {
-  Step = 'step',
-  Schema = 'schema',
-  ReturnUrl = 'returnUrl',
-  IsCreating = 'isCreating',
-  AppId = 'appId',
-}
-
-export enum AppsEditorStep {
-  General = 'general',
-  Settings = 'settings',
-}
-
 export const QUERY_VALUE_TRUE = '1';
 export const READY_TO_INTERACT_EVENT = 'readyToInteract';
 export const UPDATED_SUCCESS_EVENT = 'updatedApplicationSuccess';
