+++ apps/chat/src/types/apps-editor.ts
@@ -1,0 +1,12 @@
+export enum AppsEditorQuery {
+  Step = 'step',
+  Schema = 'schema',
+  ReturnUrl = 'returnUrl',
+  IsCreating = 'isCreating',
+  AppId = 'appId',
+}
+
+export enum AppsEditorStep {
+  General = 'general',
+  Settings = 'settings',
+}
